Create separate visitors for generating C++ files and computing C++ filenames
